[git-p4: depot-paths = "//open/dt/dev/": change = 5072]

+++ farrago/src/com/disruptivetech/farrago/calc/RexToCalcTranslator.java
@@ -112,11 +112,9 @@ public class RexToCalcTranslator implements RexVisitor
                 new TypePair(
                     fac.createSqlType(SqlTypeName.Bigint),
                     CalcProgramBuilder.OpType.Int8),
-                // TODO jvs 10-Oct-2005:  get rid of this once Fennel
-                // calc supports fixed-point numerics
                 new TypePair(
                     fac.createSqlType(SqlTypeName.Decimal),
-                    CalcProgramBuilder.OpType.Double),
+                    CalcProgramBuilder.OpType.Int8),
                 new TypePair(
                     fac.createSqlType(SqlTypeName.Float),
                     CalcProgramBuilder.OpType.Double),
